feat: add mood summary to monthly archive pages
- Extract EMOTIONS dictionary to shared lib/moodUtils.ts
- Add analyzeEmotions() utility for mood detection
- Refactor moods.tsx to use shared utilities (DRY)
- Display top 3 emotions in archive month stats line
